got rid of longs

+++ AE_PunktBeschriftung/include/representations/Point.hpp
@@ -10,7 +10,7 @@ class Point {
 public:
 
     using Ptr = std::shared_ptr<Point>;
-    using Point2D = std::tuple<long, long>;
+    using Point2D = std::tuple<int, int>;
     using Rectangle = std::tuple<Point2D, Point2D>;
 
     enum Corner {
@@ -20,10 +20,10 @@ public:
 private:
 
     int idx;
-    long x;
-    long y;
-    long width;
-    long height;
+    int x;
+    int y;
+    int width;
+    int height;
     std::string label;
 
     std::vector<Point::Ptr> neighbours;
@@ -32,13 +32,13 @@ private:
 
 public:
 
-    Point(int idx, long x, long y, long width, long height, std::string label);
+    Point(int idx, int x, int y, int width, int height, std::string label);
 
     int getIdx() const;
 
-    long getX() const;
+    int getX() const;
 
-    long getY() const;
+    int getY() const;
 
     std::string getName() const;
